handle pclint warnings

+++ src/scip/implics.c
@@ -317,7 +317,7 @@ SCIP_RETCODE SCIPvboundsDel(
    if( (*vbounds)->len == 0 )
       SCIPvboundsFree(vbounds, blkmem);
 
-   return SCIP_OKAY;
+   return SCIP_OKAY; /*lint !e438*/
 }
 
 /** reduces the number of variable bounds stored in the given variable bounds data structure */
@@ -901,7 +901,7 @@ void SCIPimplicsGetVarImplics(
 
    *haslowerimplic = (poslower >= 0);
    *hasupperimplic = (posupper >= 0);
-}
+}  /*lint !e438*/
 
 /** returns whether an implication y <= b or y >= b is contained in implications for x == 0 or x == 1 */
 SCIP_Bool SCIPimplicsContainsImpl(
@@ -3367,13 +3367,13 @@ unsigned int SCIPcliqueGetId(
    SCIP_CLIQUE*          clique              /**< clique data structure */
    )
 {
-   int id;
+   unsigned int id;
 
    assert(clique != NULL);
 
    id = clique->id;
 
-   return (unsigned int)id;
+   return id;
 }
 
 /** gets index of the clique in the clique table */
